% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_shapes_module.R
\name{complete_shapes_module}
\alias{complete_shapes_module}
\alias{complete_shapes_ui}
\alias{complete_shapes_server}
\title{Complete Halved Shapes Module}
\usage{
complete_shapes_ui(id)

complete_shapes_server(id)
}
\arguments{
\item{id}{Module id}
}
\value{
Invisibly returns a reactive containing the last results
}
\description{
UI and server logic to complete halved shapes into symmetrical images
using the `complete_halved_shape()` function with batch processing support.
}
