% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_saturation.R
\name{compute_pca_saturation}
\alias{compute_pca_saturation}
\title{Compute PCA Saturation Curve}
\usage{
compute_pca_saturation(
  pca_object,
  sample_sizes = seq(0.1, 1, 0.1),
  bootstrap_iterations = 200,
  pcs_to_analyze = NULL,
  metric = c("total_variance", "cumulative_variance"),
  min_sample_size = 5,
  seed = NULL,
  parallel = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{pca_object}{PCA object or matrix of PC scores}

\item{sample_sizes}{Sample sizes to test}

\item{bootstrap_iterations}{Number of bootstrap iterations}

\item{pcs_to_analyze}{Number of PCs to analyze}

\item{metric}{Variance metric to compute}

\item{min_sample_size}{Minimum sample size}

\item{seed}{Random seed}

\item{parallel}{Use parallel processing}

\item{n_cores}{Number of cores}
}
\value{
List with saturation data and parameters
}
\description{
Compute PCA Saturation Curve
}
