% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_detection.R
\name{.create_analysis_grid}
\alias{.create_analysis_grid}
\title{Create Analysis Grid}
\usage{
.create_analysis_grid(x_range, y_range, resolution, buffer)
}
\description{
Create Analysis Grid
}
\keyword{internal}
