% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_detection.R
\name{.create_domain_hull}
\alias{.create_domain_hull}
\title{Create Domain Hull}
\usage{
.create_domain_hull(points, hull_type, alpha_value, buffer, verbose)
}
\description{
Create Domain Hull
}
\keyword{internal}
