% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_detection.R
\name{.extract_gap_polygons}
\alias{.extract_gap_polygons}
\title{Extract Gap Polygons from Certainty Matrix}
\usage{
.extract_gap_polygons(
  gap_certainty,
  grid_x,
  grid_y,
  in_domain,
  certainty_thresholds,
  domain_hull,
  points
)
}
\description{
Extract Gap Polygons from Certainty Matrix
}
\keyword{internal}
