% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbow_plot.R
\name{elbow_plot}
\alias{elbow_plot}
\title{Elbow Plot for K-means Clustering}
\usage{
elbow_plot(
  data,
  x_col,
  y_col,
  max_k = 10,
  scale = TRUE,
  title = "Elbow Plot",
  x_label = "Number of Clusters (k)",
  y_label = "Total Within-Cluster Sum of Squares",
  export = FALSE,
  file_name = "elbow_plot",
  file_path = NULL
)
}
\arguments{
\item{data}{A data frame containing the points to cluster.}

\item{x_col}{A character string specifying the column name for the x-axis.}

\item{y_col}{A character string specifying the column name for the y-axis.}

\item{max_k}{An integer specifying the maximum number of clusters to consider. Default is 10.}

\item{scale}{Logical. Whether to scale the \verb{(x, y)} points before clustering. Default is TRUE.}

\item{title}{A character string specifying the title of the plot. Default is "Elbow Plot".}

\item{x_label}{A character string specifying the label for the x-axis. Default is "Number of Clusters (k)".}

\item{y_label}{A character string specifying the label for the y-axis. Default is "Total Within-Cluster Sum of Squares".}

\item{export}{Logical. Whether to export the plot as a TIFF file. Default is FALSE.}

\item{file_name}{A character string specifying the file name if \code{export = TRUE}. Default is "elbow_plot".}

\item{file_path}{An optional path for saving the exported file. If NULL, the file is saved in the working directory.}
}
\value{
A ggplot2 object visualizing the elbow plot.
}
\description{
This function visualizes the elbow point for k-means clustering to help determine
the optimal number of clusters. It calculates the total within-cluster sum of squares (WCSS)
for a range of cluster numbers and plots the WCSS against the number of clusters.
}
