% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hull_specimen_table.R
\name{get_hull_specimen_table}
\alias{get_hull_specimen_table}
\title{Get Combined Specimen Data Table for Convex Hulls of Multiple Column Pairs}
\usage{
get_hull_specimen_table(
  data,
  x_cols,
  y_cols,
  group_col,
  group_vals = NULL,
  max_rows_per_page = 50
)
}
\arguments{
\item{data}{A data frame containing the data to analyze.}

\item{x_cols}{A character vector representing the x-axis column names.}

\item{y_cols}{A character vector representing the y-axis column names.}

\item{group_col}{A character string representing the grouping column name.}

\item{group_vals}{(Optional) A vector of specific values within the group_col to filter by for hull calculation.
If NULL, tables are generated for all groups in \code{group_col}.}

\item{max_rows_per_page}{Maximum number of rows per page before splitting. Default is 50.}
}
\value{
A list of ggplot tables, each displaying a portion of the combined data,
with titles as bold rows and automatic pagination if necessary.
}
\description{
This function identifies the specimens from a data frame used to create convex hulls
based on multiple specified column pairs for specific groups and combines them into a single table.
Each group and column pair has a title row with the format "Specimens for hull of group: ... for X vs Y".
}
