% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_shapes_to_data.R
\name{map_shapes_to_data}
\alias{map_shapes_to_data}
\title{Map Shapes to Data}
\usage{
map_shapes_to_data(data, id_col, shape_folder)
}
\arguments{
\item{data}{A data frame containing the IDs used for matching with shape file names.}

\item{id_col}{A character string specifying the column in the data frame that contains IDs to be matched with the shape file names.}

\item{shape_folder}{A character string specifying the folder containing the \code{.jpeg} shape files.}
}
\value{
A data frame with an additional \code{shape} column containing imported shapes as \code{Out} objects or \code{NULL} for unmatched rows.
}
\description{
This function maps shapes from a folder of \code{.jpeg} files to a specified column in a data frame.
The shape files are imported and processed into \code{Out} objects using the \code{Momocs} package.
If a match between the file names (excluding the \code{.jpeg} extension) and the values in the specified ID column is found,
the corresponding shape is added to the data frame in a new column named \code{shape}.
Otherwise, \code{NULL} is assigned to rows without a match.
}
