% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_module.R
\name{overview_ui}
\alias{overview_ui}
\title{Overview Module (Haug_overview)}
\usage{
overview_ui(id)
}
\arguments{
\item{id}{Module id}

\item{data_reactive}{A reactive function returning a data.frame from Data Import}
}
\description{
UI and server to configure and render the Haug_overview() results on the
imported dataset. Users select an even number of columns (paired as 1-2,
3-4, ...), optional grouping, and basic options for hulls/contours/export.
}
