% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_saturation_module.R
\name{pca_saturation_server}
\alias{pca_saturation_server}
\title{PCA Saturation Curve Module - Server}
\usage{
pca_saturation_server(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
Reactive values with analysis results
}
\description{
Shiny module server for PCA saturation curve analysis
}
