% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_saturation_module.R
\name{pca_saturation_ui}
\alias{pca_saturation_ui}
\title{PCA Saturation Curve Module - UI}
\usage{
pca_saturation_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
Shiny UI elements
}
\description{
Shiny module UI for PCA saturation curve analysis
}
