% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_plot.R
\name{print.cluster_plot_result}
\alias{print.cluster_plot_result}
\title{Print method for cluster_plot_result objects}
\usage{
\method{print}{cluster_plot_result}(x, ...)
}
\arguments{
\item{x}{A cluster_plot_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for cluster_plot_result objects
}
