% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_halved_shape.R
\name{print.complete_shape_result}
\alias{print.complete_shape_result}
\title{Print method for complete_shape_result objects}
\usage{
\method{print}{complete_shape_result}(x, ...)
}
\arguments{
\item{x}{A complete_shape_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for complete_shape_result objects
}
