% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_image_right_half.R
\name{print.cut_image_result}
\alias{print.cut_image_result}
\title{Print method for cut_image_result objects}
\usage{
\method{print}{cut_image_result}(x, ...)
}
\arguments{
\item{x}{A cut_image_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for cut_image_result objects
}
