% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbow_plot.R
\name{print.elbow_plot_result}
\alias{print.elbow_plot_result}
\title{Print method for elbow_plot_result objects}
\usage{
\method{print}{elbow_plot_result}(x, ...)
}
\arguments{
\item{x}{An elbow_plot_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for elbow_plot_result objects
}
