% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haug_overview.R
\name{print.haug_overview_result}
\alias{print.haug_overview_result}
\title{Print method for haug_overview_result objects}
\usage{
\method{print}{haug_overview_result}(x, ...)
}
\arguments{
\item{x}{A haug_overview_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for haug_overview_result objects
}
