% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haug_panel.R
\name{print.haug_panel_result}
\alias{print.haug_panel_result}
\title{Print method for haug_panel_result objects}
\usage{
\method{print}{haug_panel_result}(x, ...)
}
\arguments{
\item{x}{A haug_panel_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for haug_panel_result objects
}
