% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hull_specimen_table.R
\name{print.hull_specimen_table_result}
\alias{print.hull_specimen_table_result}
\title{Print method for hull_specimen_table_result objects}
\usage{
\method{print}{hull_specimen_table_result}(x, ...)
}
\arguments{
\item{x}{A hull_specimen_table_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for hull_specimen_table_result objects
}
