% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_plot_momocs.R
\name{print.momocs_plot_result}
\alias{print.momocs_plot_result}
\title{Print method for momocs_plot_result objects}
\usage{
\method{print}{momocs_plot_result}(x, ...)
}
\arguments{
\item{x}{A momocs_plot_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for momocs_plot_result objects
}
