% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_shapes.R
\name{print.morph_shapes_result}
\alias{print.morph_shapes_result}
\title{Print method for morph_shapes_result objects}
\usage{
\method{print}{morph_shapes_result}(x, ...)
}
\arguments{
\item{x}{A morph_shapes_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for morph_shapes_result objects
}
