% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_detection.R
\name{print.morphospace_gaps}
\alias{print.morphospace_gaps}
\title{Print Method for morphospace_gaps}
\usage{
\method{print}{morphospace_gaps}(x, ...)
}
\arguments{
\item{x}{Object of class morphospace_gaps}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for morphospace_gaps
}
