% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_saturation.R
\name{print.pca_saturation}
\alias{print.pca_saturation}
\title{Print method for pca_saturation objects}
\usage{
\method{print}{pca_saturation}(x, ...)
}
\arguments{
\item{x}{A pca_saturation object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for pca_saturation objects
}
