% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_analysis.R
\name{print.shape_analysis_result}
\alias{print.shape_analysis_result}
\title{Print method for shape_analysis_result objects}
\usage{
\method{print}{shape_analysis_result}(x, ...)
}
\arguments{
\item{x}{A shape_analysis_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for shape_analysis_result objects
}
