% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_shapes_to_data.R
\name{print.shape_mapped_data}
\alias{print.shape_mapped_data}
\title{Print method for shape_mapped_data objects}
\usage{
\method{print}{shape_mapped_data}(x, ...)
}
\arguments{
\item{x}{A data frame with shape mapping results}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for shape_mapped_data objects
}
