% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_image.R
\name{print.split_image_result}
\alias{print.split_image_result}
\title{Print method for split_image_result objects}
\usage{
\method{print}{split_image_result}(x, ...)
}
\arguments{
\item{x}{A split_image_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for split_image_result objects
}
