% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momocs_cl-helpers.R
\name{verify}
\alias{verify}
\title{Validates Coo objects}
\usage{
verify(Coo)
}
\arguments{
\item{Coo}{any Coo object}
}
\value{
a Coo object.
}
\description{
No validation for S3 objects, so this method is a (cheap) attempt at checking
Coo objects, Out, Opn and Ldk objects.
}
