% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_saturation.R
\name{plot_pca_saturation}
\alias{plot_pca_saturation}
\title{Plot PCA Saturation Curve}
\usage{
plot_pca_saturation(
  saturation_result,
  x_axis = c("absolute", "proportion"),
  normalize_metrics = TRUE,
  show_ci = TRUE,
  show_points = TRUE,
  colors = NULL,
  theme_name = "Haug",
  title = "PCA Saturation Curve",
  subtitle = NULL
)
}
\arguments{
\item{saturation_result}{Output from compute_pca_saturation}

\item{x_axis}{Type of x-axis}

\item{normalize_metrics}{Normalize metrics to percentage}

\item{show_ci}{Show confidence intervals}

\item{show_points}{Show individual data points}

\item{colors}{Color vector for metrics}

\item{theme_name}{Plot theme}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}
}
\value{
A ggplot2 object
}
\description{
Plot PCA Saturation Curve
}
